+++ NEWS.md
@@ -10,6 +10,10 @@
 ## New features
 - preliminary integration of code from `SCnext/mixedSCGLR` written by Jocelyn Chauvet (issue #11)
 
+## Deprecations
+- parameter `nfolds` of `scglrCrossVal` as been renamed to `folds` and now accepts also vectors.
+For now a warning is issued and provided value is used but update your code as it will be removed
+in a next version.
 
 # version 3.0
 This major version introduces a new feature allowing to group covariates in so called **themes**.